# CTHz/openapi.yaml
openapi: 3.0.3
info:
  title: CTHz System 300 Device Management API (Internal)
  version: 0.1.0
  description: |
    Internal-only API specification for the System 300 Device Management (DM) service.

    Notes
    - This spec includes Spurtree-owned Device Management APIs and references CTHz-owned Sensor APIs (proxied by DM).
    - Do not expose Swagger/Redoc in production. Keep this file internal.
servers:
  - url: https://localhost:8443
    description: Local development (TLS)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Device:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        device_id: { type: string }
        serial_number: { type: string }
        ip_address: { type: string }
        status: { type: string, example: online }
    Station:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        device_ids: { type: array, items: { type: string } }
...
paths:
...
  /api/devices:
    get:
      tags: [devices]
      summary: List managed devices
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      responses:
        '200': { description: Devices, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Device' } } } } }
    post:
      tags: [devices]
      summary: Add device manually
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Device' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Device' } } } }
  /api/devices/{id}:
    get:
      tags: [devices]
      summary: Get device details
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Device, content: { application/json: { schema: { $ref: '#/components/schemas/Device' } } } }
        '404': { description: Not found }
    put:
      tags: [devices]
      summary: Update device info
      x-owner: Spurtree
      x-priority: Medium
      x-phase: MVP
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Device' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Device' } } } }
    delete:
      tags: [devices]
      summary: Remove device
      x-owner: Spurtree
      x-priority: Medium
      x-phase: MVP
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
...
  /api/stations/{id}/devices:
    post:
      tags: [stations]
      summary: Add device to station
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device_id: { type: string }
      responses:
        '200': { description: Added, content: { application/json: { schema: { $ref: '#/components/schemas/Station' } } } }
  /api/stations/{id}/devices/{device_id}:
    delete:
      tags: [stations]
      summary: Remove device from station
      x-owner: Spurtree
      x-priority: Medium
      x-phase: MVP
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: device_id, required: true, schema: { type: string } }
      responses:
        '200': { description: Removed, content: { application/json: { schema: { $ref: '#/components/schemas/Station' } } } }

  # ==========================
  # Policies (Spurtree-owned)
  # ==========================
...
  /api/config/devices/{id}:
    get:
      tags: [config]
      summary: Get all device configuration
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses:
        '200': { description: Config, content: { application/json: { schema: { type: object } } } }
    put:
      tags: [config]
      summary: Update device configuration
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { type: object } } } }
    patch:
      tags: [config]
      summary: Partial configuration update
      x-owner: Spurtree
      x-priority: Medium
      x-phase: MVP
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { type: object } } } }
...
  /api/monitoring/devices/{id}:
    get:
      tags: [monitoring]
      summary: Individual device status
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses:
        '200': { description: Status, content: { application/json: { schema: { type: object } } } }
...
  /scans/{scan_id}:
    get:
      tags: [sensor]
      summary: Get scan results by ID
      x-owner: CTHz
      x-priority: High
      x-phase: MVP
      parameters: [{ in: path, name: scan_id, required: true, schema: { type: string } }]
      responses:
        '200': { description: Scan, content: { application/json: { schema: { type: object } } } }
...
  /calibration/{calibration_id}:
    get:
      tags: [sensor]
      summary: Get calibration results
      x-owner: CTHz
      x-priority: Medium
      x-phase: MVP
      parameters: [{ in: path, name: calibration_id, required: true, schema: { type: string } }]
      responses:
        '200': { description: Calibration, content: { application/json: { schema: { type: object } } } }
...
  /config/device/info:
    put:
      tags: [sensor]
      summary: Basic device info (CFG-001)
      x-owner: CTHz
      x-priority: High
      x-phase: MVP
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { type: object } } } }
...